🦄 refactor(fs): 重构目录项相关代码

+++ os/src/fs/fat32/super_block.rs
@@ -26,8 +26,8 @@ impl Fat32SB {
 }
 
 impl Fat32SB {
-    pub fn root_sector(&self) -> u32 {
-        let res = self.reserved_sectors_cnt as u32 + self.fat_cnt as u32 * self.fat_size_32;
+    pub fn root_sector(&self) -> usize {
+        let res = self.reserved_sectors_cnt as usize + self.fat_cnt as usize * self.fat_size_32 as usize;
         res
     }
 }
